+++ framework/model/src/main/java/is/codion/framework/model/EntityModel.java
@@ -103,7 +103,7 @@ public interface EntityModel<M extends EntityModel<M, E, T>, E extends EntityEdi
 		 * @return detail models with an active link to this model, that is, those that should respond to master model events
 		 * @see ModelLink#active()
 		 */
-		ObservableValueSet<M> linked();
+		ObservableValueSet<M> active();
 
 		/**
 		 * Adds the given detail model to this model, a side effect if the detail model contains